chore(core): gather memory files
